requirements.txt
chromadb
sentence-transformers
google-generativeai
python-dotenv
SpeechRecognition
PyAudio
pyttsx3
gTTS
playsound==1.2.2
Flask
transformers>=4.30.0
torch>=2.0.0
numpy
google-generativeai>=0.7.2
